requirements/prod.txt
-r base.txt

gunicorn==19.9.0
eventlet==0.25.1
django-storages[azure]==1.10.1
